+++ cocos/renderer/backend/CommandBuffer.h
@@ -21,26 +21,26 @@ public:
     virtual void beginFrame() = 0;
     virtual void beginRenderPass(const RenderPassDescriptor& descriptor) = 0;
     virtual void setRenderPipeline(RenderPipeline* renderPipeline) = 0;
-    virtual void setViewport(ssize_t x, ssize_t y, size_t w, size_t h) = 0;
+    virtual void setViewport(int x, int y, unsigned int w, unsigned int h) = 0;
     virtual void setCullMode(CullMode mode) = 0;
-    virtual void setVertexBuffer(size_t index, Buffer* buffer) = 0;
+    virtual void setVertexBuffer(unsigned int index, Buffer* buffer) = 0;
     virtual void setBindGroup(BindGroup* bindGroup) = 0;
     virtual void setIndexBuffer(Buffer* buffer) = 0;
-    virtual void drawArrays(PrimitiveType primitiveType, uint32_t start,  uint32_t count) = 0;
-    virtual void drawElements(PrimitiveType primitiveType, IndexFormat indexType, uint32_t count, uint32_t offset) = 0;
+    virtual void drawArrays(PrimitiveType primitiveType, unsigned int start,  unsigned int count) = 0;
+    virtual void drawElements(PrimitiveType primitiveType, IndexFormat indexType, unsigned int count, unsigned int offset) = 0;
     virtual void endRenderPass() = 0;
     virtual void endFrame() = 0;
     
     virtual void setLineWidth(float lineWidth) = 0;
     
-    void setStencilReferenceValue(uint32_t value);
-    void setStencilReferenceValue(uint32_t frontRef, uint32_t backRef);
+    void setStencilReferenceValue(unsigned int value);
+    void setStencilReferenceValue(unsigned int frontRef, unsigned int backRef);
     
 protected:
     virtual ~CommandBuffer() = default;
     
-    uint32_t _stencilReferenceValueFront = 0;
-    uint32_t _stencilReferenceValueBack = 0;
+    unsigned int _stencilReferenceValueFront = 0;
+    unsigned int _stencilReferenceValueBack = 0;
 };
 
 CC_BACKEND_END
